`au` is local, and only happens "once" (`au!`)
The way this was being defined globally was fine w.r.t. the pattern of
buffers it matches; however, it would define a new global `autocmd`
every time this ftplugin is loaded.

Which is not optimal.

Here we restrict it to the specific buffer only, and past that, cause
the `autocmd` to overwrite itself if the ftplugin is invoked again.